// src/app/service consumer app/model/serviceCounsumerApp.ts
export class ServiceConsumerApp {
    id: number;
    name: string;
    description: string;
    key: string;
    password: string;
    appCategoryId: number;
    appCategoryName: string;
    token: string;
    tokenExpiryDateTime: string;
    status:string;
    createdBy:number;
    updatedBy:number;

}